FIX: Stop double file dialog on mobile preventing uploads
There is a separate mobile-file-upload button used for the
composer that is bound with `_bindMobileUploadButton`, which
shows the file selection dialog box when clicked. However this
is leading to a double file dialog on mobile for the chat-composer,
because there is no concept of a mobile upload button there.

This commit just changes `_bindMobileUploadButton` to a noop for
the chat-composer to prevent the issue from happening.