UX: Increase the time before a user is marked 'offline'
The PresenceChannel default is 'browser in the background for 10s', which is a little aggressive for this use case. Users appear to constantly go online/offline, which is quite annoying. This commit updates the timeout to 60 seconds.

Requires https://github.com/discourse/discourse/pull/15217 to make it work, but won't blow up if installed on an earlier Discourse version