FEATURE: Slack-compatible webhooks (#289)
This commit adds support for slack-compatible webhooks, which
many other integrations already support (including Mattermost,
OpsGenie, etc.)

To use the slack compatible version of the webhook, post to
the webhook URL and append `/slack` at the end, e.g.

https://meta.discourse.org/chat/hooks/ID/slack.json

See https://api.slack.com/reference/messaging/payload for the
slack message payload format. For now we only support the
text param, which we preprocess lightly to remove the slack-isms
in the formatting.

Later on down the track we may want to process blocks, attachments,
and user mentions in the text.