FEATURE: Handle attachments for Slack webhooks (#291)
Very minimal handling of the attachments part of the
slack webhook payload, to handle our OpsGenie integration,
which does not use the top-level `text` field.

This is quite barebones and needs to be expanded upon in
future. We are just using the `fallback` field for now because
it contains all the text we need from the other attachment
fields. See https://api.slack.com/reference/messaging/attachments
for more details.

Also added a SiteSetting.chat_debug_webhook_payloads setting
to drop the webhook payloads into the Rails logger when
enabled, so we can look at them in a nicer way.