FIX: Don't scroll the entire chat pane to GitHub oneboxes (#329)
When a GitHub onebox renders off-screen, calling `scrollIntoView` would scroll the entire chat window up to the onebox. This is not what we want.

All we want to do is scroll the `.onebox .onebox-body` div so that the selected line is in view

Unfortunately calculating the y-position of that line relative to the `.onebox-body` is quite complicated, but I think this should be pretty robust.