UX: remove online count from channels
Online count is a bit rough at the moment due to online definition.

It tends to go up and down, no way to easily drill in and see who and just
is becoming somewhat distracting and confusting.

You can pick the people you would like to see online status of by adding to
personal chat.

This has better parity with other chat application that do not surface
the total number.