FEATURE: Expanded drop target for files
The entire .tc-live-pane element can now be used to
drop files into for uploads. An .uppy-is-drag-over class
is added to the element on drag over and removed on
drag leave, and we can also hook into those events
if we want to do anything fancy with CSS or JS. This
commit just expands the area so it's no longer just
the text box.