Use ember component for collapsing instead of yanking into a container
A few bugs were occurring:
- When editing the youtube URL to something else, the lazyYT invocation
was causing Ember to error out when clearing child nodes, because of the
way collapsing was done. This caused the edit to not show up properly
- The title kept getting added when a new message appeared in the chat
because we were redecorating every message very often.

Using the ember component will make sure it renders nicely with the
lifecycle of the message.

and maybe also we can reuse it for images?