FIX: ensures chat icon is not shown for closed topics
Not everything is tested as I tried to position the check as soon as I could but in any case the widget behaviour is fully tested and will be hidden if necessary. So worst case would be we add to the DOM a widget we could have avoided but it will be hidden.