FIX: Finally max-width and max-height images using object fit (#518)
This was tested also with twitter image embeds minimally. Maybe I'd get this right for reals. This was done with `object-fit: contain`. `width` needs to be `unset` because of the way our onebox engines are setting certain images.

One thing we may consider is a site setting for max-height. It's currently hardcoded to 150px.

Note that this fix cannot apply to youtube and we would have to change core because of the way lazyYT is dynamically calculating positions for the buttons and such.