FIX: Only support a subset of markdown. (#468)
Restrict markdown for chat to a subset of features which must be
explictly enabled. The previous pattern which we relied on required us
to selectively opt out. However, anything new added to core will
automatically be made available until it is opted out in this plugin. If
there is any delay in between, this means that `ChatMessage#cooked` can
differ between messages even though `ChatMessage#cooked_version` remains
the same. By switching to this new pattern, we ensure that the way we
cook a message is always the same given a similar
`ChatMessage#cooked_version`.

This commit also fixes a bug where we were trying to opt out of
Markdown-It rules via the `feaures` options which actually doesn't work.

cf. c2afc3915b8d533858849c7973de6e1038c81245