FIX: Show lightbox after collapse and expand (#528)
The lightboxing function is applied once when messages are loaded. Since collapsing an image removed it from the document entirely, the lightboxing feature wasn't working after a user collapses and expands the image.

To prevent this we could either A: reapply the lightboxing function every time after expand, or B: just skip removing the image when it's collapsed and just hide it. I chose B.

Note on test:
`click(".mfp-container")` was the original but it didn't work I suspect due to how the overlay is rendered in tests. To find the container we need to use the document.