DEV: Do not lookup services in test beforeEach hook (#561)
This causes the services to be initialized before the app has started. This means that the inject-discourse-objects pre-initializer has not run, and so the service will be initialized without any of the standard injections (e.g. currentUser, siteSettings, etc.)

This commit updates the hooks to define getters which will lookup the services 'just in time'.