DEV: Restore test runloop hack for legacy ember environment
a897f86a removed the use of `next()` in this plugin's tests, so that they passed consistently on Ember CLI. Unfortunately that change broke the tests under the legacy Ember environment. This commit restores the hack **for the legacy environment only**, and wraps it up in a `chatSettled` helper so we can easily remove it in future.