FIX: Preload custom fields after server reload (#573)
Preload after reload, what a concept! 😛

I believe this fixes the issue I was having in dev env, where after a server reload the app would complain that some custom fields were not preloaded. This was happening because another plugin was preloading them after a reload, which enables the "warn on non-preloaded" logic, and those cfs from chat were missing.