FEATURE: Chat transcripts using [chat] custom markdown BBCode (#544)
This commit introduces the new `[chat]` BBCode block, which will be used for quoting chat messages into regular posts. The chat messages quoted inside `[chat]` will be rendered with the same limited subset of BBCode defined in `ChatMessage::MARKDOWN_FEATURES` and `ChatMessage::MARKDOWN_IT_RULES` as the server-side chat messages are cooked with. This is done via the new `requestCustomMarkdownCookFunction` API in core.

The `[chat]` block requires a `quote` attribute in the following format: `quote="username;message_id;datetime"`. It allows for an optional `channel` attribute which should be the channel name, and this should only be used sparingly. Here is a complete example:

```
[chat quote="martin;2345;2022-01-25T05:40:39Z" channel="Lounge"]
This is my cool message!
[chat]
```

This is intended to be used as a single message per `[chat]` block, except for in the case of multiple subsequent messages from the same user. Most of the time this BBCode will be autogenerated, so the ergonomics of how it can be typed in are of a lower priority.

The formatted datetime is rendered using the `decorateCookedElement` plugin API for each post.