FEATURE: Add multiquote and chained attrs for chat-transcript (#594)
The multiQuote attribute for [chat] makes the channel name show
above the username + datetime. If the channel name is provided
and the multiQuote attribute is not present or false, then the
channel name is put to the right of the datetime of the chat message.
This commit also links to the chat channel by name and the corresponding
ember route.

Also in this commit is the chained CSS class that can be added
to `[chat]`. This is used on subsequent messages from the same
channel so no margin is added to the chat messages, making them
all stick together with CSS. This is added via a new `chained` attribute
on the BBCode.