FIX: User could not re-enable chat after disabling it (#595)
In 18f45fc we made a change to not
call addSaveableUserOptionField for the chat-specific user options
if !userCanChat. However, this prevents the user from changing their
chat preferences if they disable chat and press save. They are not able
to re-enable it or change any other settings, because the user options
are no longer added.