FEATURE: Chat quoting (#601)
This commit introduces chat message quoting functionality. Selected messages from any channel _except_ DM channels can be quoted.

This commit also allows non-staff users to bulk select messages, because it is necessary for quoting. The "Move to Topic" button is hidden for non-staff.

Quoting has a few different behaviours based on the current state:

1. When isolated on desktop, copy to clipboard and show a little message
2. Not isolated on desktop, open composer and insert markdown. This may open a reply composer if the user is currently inside a topic, otherwise it will open a new topic composer
3. On mobile, go to the category / topic that the chat is for and open the composer and insert markdown

When the quote button is clicked, a call to the `/chat/:channelId/quote` endpoint is sent with the selected message IDs. The quote BBCode is built with `ChatTranscriptService`. The bbcode is constructed server-side because other functions, such as Move to Topic and Archive Channel will need to build up the same bbcode later on. See the comment on `ChatTranscriptService` for further information on how that works.