FIX: Don't create errored chat message from webhook event (#621)
When we create chat messages with ChatMessageCreator, there
is a chance the message will fail and return an error message.
This was working fine, however when we were creating the
ChatWebhookEvent record we were passing the `@chat_message` to
it, which meant that the chat message was always created regardless
of whether the chat message creator errored or not.

This commit fixes the issue by only making the ChatWebhookEvent once
the chat message has been validated. Any other validations must
occur before the chat message is saved, and any records created
linking to the message must be created after the message is saved.