FEATURE: Chat channel statuses (#619)
This PR adds `status` to chat channel, with `open`, `closed`, `read_only`, and `archived` statuses (default `open`).

Rules have been introduced in the UI and in the backend to handle the various channel statuses, with the `Guardian` extensions focused on, with a lot of new tests there. No UI exists yet to change the channel status, but only staff will be able to do that in future. When the channel status is changed, we log it in the staff action log, fire a `DiscourseEvent`, and a message bus message as well.

The rules for each status are, in brief. If `open` then anyone can post in the channel, edit messages, and delete messages.  If `closed` then only staff can do those things. If `archived` or `read_only` noone can do those things.

Also some improvements to this PR in the chat JS. I've now started using the `ChatChannel` JS model so we can start adding shortcut functions for it, replacing the bare `EmberObject` we were using previously. I would like to make another PR later to avoid things like this, and have `.isDirectMessageChannel()` instead:

```javascript
const inDmChannel = activeChannel.chatable_type === CHATABLE_TYPES.directMessageChannel;
```