FIX: New DM erroring on create because of canModifyMessages (#643)
When we were creating a new DM channel in non-isolated chat
mode, we were not creating a ChatChannel instance to wrap
the returned channel in, meaning the canModifyMessages
function cannot be found and throws an error. This is not
a problem in isolated chat because the channel is loaded via
a route.