FEATURE: Chat channel archiving (#651)
This PR introduces the ability for staff to archive chat channels. The flow looks like this:

1. The admin chooses "Archive Channel" from the new settings dropdown
2. The admin selects either a new topic or an existing topic to archive the messages in the channel to
3. We create a `ChatChannelArchive` record in the database to store their selection and the user doing the archiving, and we change the channel status to `read_only` so no messages in the channel can be added or changed
4. We run a background job that creates a post for each 100 messages using the `[chat]` quoting markdown, as well as creating the new topic if that was required. We trash the messages in each batch before proceeding to the next batch.
5. We send a PM to the user who initiated the archive on completion

The job is designed to be idempotent. If the job fails the admin is sent a PM prompting them to retry. We log errors in the background. A follow up PR will add the retry archive button in the UI