FIX: Move channel settings button to channel browse list
This commit moves the setting button out of the chat sidebar
on hover and into the channel browse list, because it is not
an often used action. Also shortens the status message at the
top of the channel, and adds a status indicator for the channel
in the browse list.