FEATURE: Retry failed archive (#674)
This commit introduces the ability to retry a failed chat channel archive. This should be rare, but the job is idempotent so we are introducing a way for admins to retry an archive that fails partway through. This section now shows at the top of the channel with a failed archive.

All the retry button does is re-enqueue job. This commit also introduces a "channel-archive-status" message bus message that will automatically update the UI when the archive fails. We can also use this in future if we want to have some sort of progress bar in the UI.