FIX: Guardian can_rebake? was breaking core post rebake endpoint
In the guardian extension we added a can_rebake? method with
a chat message as an argument. This causes a 500 error in the
core /post/:id/rebake endpoint, because guardian.ensure_can_rebake!
and guardian.can_rebake? are now expecting a chat message argument.
This commit just changes the chat guardian extension to be
can_rebake_chat_message?