FEATURE: Include reactions in archive chat quotes
This commit introduces a new reactions attribute to
the [chat] BBcode in the format `emoji:user1,user2;emoji2:user1`.
When included, we store this data in a data-reactions attr
on the wrapping chat transcript div, and also construct
reaction indicators at the bottom of each [chat] quoted
set of messages.

We have intentionally left off the following:

* Fidelity is only per-group of messages. For example, if there
  are 3 messages from a single user grouped into one [chat]
  block, and each individual message has a reaction, all 3
  reactions are placed at the bottom of the set of messages.
* When hovering the reactions, we do not show which username
  reacted to what, we should be easily able to do this in
  V2 of this feature.

Regular quoting of chat messages will not include
reaction metadata, only when the messages are archived.