FIX: Use system user for archive topics and posts
When creating channel archive topics and posts previously
we used the archiving user as the creator of new topics and
posts to store the archive content. This commit changes this
to use the Discourse System user, because it is an automated
action. Also includes a few minor CSS fixes for the archive
status title.