FIX: Clipboard copy chat quotes breaks on Safari
Safari among the other browsers is doing the correct thing, and
throwing the following error when the user is clicking the
Quote button in chat isolated mode:

> NotAllowedError: the request is not allowed by the user agent or the
  platform in the current context, possibly because the user denied permission.

This is because the clipboard API is supposed to be used in a
short time period to respond to a user-initiated event. Luckily,
there is a way around this, which is to hand a promise off to
a ClipboardItem. This is done via a new clipboardCopyAsync function
created in core.

This commit uses the new function and rewrites other parts of the
chat quote code to work around its usage.