FEATURE: Remove move to topic (#720)
This feature does not exist in a form we are
happy with right now, we are removing it for
now and may come back to it later. Copying
chat messages one to one to posts is not what
we want. The alternative after this is
removed is to use chat quotes instead.