FIX: Error when anon viewing another user when chat is enabled
In 7b21a51120888caae9902c67536aaf2d78dd6964 we added a UserCardSerializer
extension for can_chat_user, which is used to show a button to initiate
a chat from the user card. However this is causing errors for anon
users viewing other users when chat is enabled, since scope.user
is nil for anon users.

This commit fixes the issue and adds tests exercising the can_chat_user
extension for the UserCardSerializer