DEV: Extract uploads from chat-composer into a chat-composer-uploads component (#764)
This commit breaks up some of the tight coupling between the chat-composer component and uploads. It removes the use of `ComposerUppyUpload` which is more complex and heavily geared towards interactions with the core composer, including the inserting of upload placeholder text which we don't need at all for the chat composer. Instead we are using `UppyUploadMixin`, which needed some improvements made in https://github.com/discourse/discourse/pull/16383 to accommodate this change.

We cannot completely remove all coupling. The chat-composer needs to load in drafts and edit messages that may have uploads attached, and for that there is an app event introduced here called `chat-composer:load-uploads` that handles this preloading. This works the other way as well -- when uploads are complete in chat-composer-uploads or removed the chat-composer needs its internal state updated, for which I introduced an `uploadChanged` action.

The chat-composer-uploads component needs to bind the paste event for the chat-composer text area to allow for uploads to be pasted in directly.

Additionally, the uploads now shows "Processing..." instead of "Uploading..." when the preprocessing such as media optimization is still happening for each individual upload. Upload state is cleared up and cancelled properly if switching to editing a message while an upload is in progress.

The CSS for the chat-composer-upload(s) has also been extracted, and some improvements have been made to e.g. allow for scrolling on horizontal overflow.