DEV: Fix composer upload event and drop target issues
This commit fixes an issue where the "Uploading..." placeholder
markdown from the composer was inserted into the chat composer.
This was due to the insert-text app event using composerEventPrefix,
which was not removed when it should have been in the
f0e777d07800dda1a0265bd57768edc4c6d8291a commit.

Also fixes an issue where the drop target was the entire page
for uploads in the floated chat, rather than just the floated
chat drawer.