FIX: Don't hide the new messages indicator beside the channel header.
Previously, seeing new messages in a channel was unnatural when there was a large number on unread messages because we displayed the newest ones first, meaning you'll have to scroll back to see the conversation from the last you
read.

This change improves our scrolling logic by loading new messages from where you left the last time you visited the channel. Also, it'll load a couple of older messages to avoid leaving the "New Messages" indicator hidden beside the
channel header. We had to refactor the messages endpoint to make the direction you're scrolling explicit, which let us create the concept of landing in the middle of the conversation, where you can generally scroll both ways.