FEATURE: Move chat messages to another channel (#804)
This commit introduces functionality for staff to move
messages from one chat channel to another. The original
messages are deleted and all references to the original
messages are updated to point to the new messages.

After the move we add a placeholder message to the original
channel indicating who moved the messages, how many were
moved, and where to, with a link to the first message in
the new channel.