DEV: Allow category group mods to moderate chat channels (#837)
This allows category group moderators to moderate chat
channels via the `guardian.can_moderate_chat?` extension.
Practically this now allows category moderators to do
the following on category channels:

* Move messages to another channel
* Delete/restore/and see deleted chat messages

It also fixes some inconsistency with `can_delete_self` and
`can_delete_others`, which we had guardian methods for
and meta/details hash to attach the result to, but we were
setting a hardcoded result in the JS code.

Doing this necessitated introducing a dynamic `generateChatView`
function for the chat fixtures, because what is in these meta properties
is dependent on the current user. Tests were updated to use this
new function instead of the static `chatView`.