FIX: Correctly scroll to the last read message.
We scroll to a previous message to avoid hiding the new message separator besides the channel header. The problem with this approach is the separator could get pushed too far down depending on the previous messages' lenght, so it
doesn't always work as expected. Instead, we'll calculate an offset using the live pane height and scroll directly to the separator, available on the next loop.