REFACTOR: ChatNotifier mention types expansion. (#842)
* REFACTOR: ChatNotifier mention types expansion.

When we want to notify users of a chat message, the `ChatNotifier` is our only source of truth. This service selects and groups the users by mention types, expanding direct, group, and channel-wide mentions. It also filters
unreachable users we can't notify because they don't have chat access or are not participating if the channel is private.

 It doesn't send notifications itself but generates a "map" and gives it to a scheduled job. This structure includes all the users to notify and the mention type metadata, which we need in order to ship different data in the
notifications.

The current implementation has a bug that includes group members who aren't participating in a private conversation if the message mentions a group. The scheduled job has an additional membership check, so it doesn't cause any
problems right now, but if another component uses this map to do other things, there's a risk it will leak private messages.

It happens because the application of the expansion rules and the map generation are happening in two different points, and we don't apply the `can_see_chat_channel?` filter during the later. The duplication makes the  logic more
error-prone, so this PR refactor the class to do both in a single step, avoiding losing information or missing filters. It also adds some unit tests to complement the acceptance ones we already have.

* FIX: Use the correct identifier when creating mention notifications.

The refactor to `CreateChatMentionNotifications` introduced a regression where the identifier wasn't using "@" for channel-wide mention. This change also includes the group name for group mention desktop and push notifications. Adds missing tests.