FIX: Excerpts for complex messages
This commit makes it so more complicated messages that
have cooked content that PrettyText.excerpt returns ""
for will just return "" for the excerpt rather than
showing raw markdown for the In Reply To message.

Also includes an improvement to just show the raw link
for the In Reply To message if it just consists of a
single URL, to get around excerpt issues with oneboxes.