FIX: Add channel ID attribute to chat quotes
We have run across various complications when trying to
link to a channel by channel title (see /t/66638 and
https://github.com/discourse/discourse-chat/pull/857 for
examples) so instead we are just going to add a channel
ID to the chat quotes instead for easier linking.

This commit also improves DM channel titles for chat quotes,
which were not taken into account in the original design,
but was later allowed in 1f10fc6c4cf5642d21419eff30dc986d0cd33a7c.