FIX: Fix filter in chat channel fetcher
This fixes the filter in the fetcher -- it was only searching
on chat channel name but that's not always there, we need to
filter on topic title and category name as well.

Also the public_channels_with_filter function was redundant,
it shared too much with secured_public_channels, so I just
moved the filter into there.