FIX: Don't hide the new messages' separator besides the channel header. (#843)
FIX: Don't hide the new messages' separator besides the channel header.

We now ensure that every time we scroll to a message, it starts at 1/6 of the chat window to avoid leaving the separator besides the channel title. 

Also, when fetching messages, if we know the last one read by the user, we'll use the lookup endpoint instead, which will let us scroll in both directions.

Finally, I included a small refactor of the `ChatController#messages` endpoint to make the direction explicit.