FEATURE: Add a button to switch back to small chat
This feature, in addition to adding a button to get back to small chat
from being in the full-page chat, adds *memory* to the window size.
This means that if you prefer to open chat in full-size when using the
sidebar it will remember which preference you have set when opening
channels.

See: https://meta.discourse.org/t/221388