PERF: Only parse HTML once during isCollapsible
Previously we were parsing message HTML 4 times for every message, and then 4 more times if the message is collapsible.

This PR updates the logic so we only parse message HTML once for each message, and then one more time if the message is collapsible.

Ideally this logic should be done during cooking on the server, but this fix will help performance for now.