FIX: Add ChatChannelFetcher specs and fix issues
Adds some more specs for ChatChannelFetcher, but there are
still some untested areas. Discovered that:

* We were checking for a muted channel when setting unread counts
  for a DM channel, but this is unnecessary because DM channels
  cannot be muted
* We were only checking if a user had a UserChatChannelMembership
  record when getting relevant DM channels, but we weren't checking
  DirectMessageUsers. Granted, it's unlikely that the former record
  would exist without the latter, but it was a little security hole.

Also changed .detect to .find...it's the same method (corporate asked
me to tell the difference between the two pictures), but in our
case .find makes more sense because we are finding a record based
on a condition.