FIX: Message order consistency and MessageMover fix
In https://meta.discourse.org/t/moving-chat-messages-to-a-new-channel-jumbled-them-all-up/227301
we were having issues with moved messages between channels
being jumbled up. During investigations it was found that
the ChatController /messages route ordered messages by ID
but everything else, including the message mover, ordered by
created_at.

This commit makes things consistent by changing the /messages
route to also order by created_at. In addition, the MessageMover
now begins moved message created_at times 5 seconds in the
future, and adds a 0.5s increment in time for each new message
created in the destination channel.

This should make it so the order is always consistent in the UI,
and so that all the timestamps of the moved messages aren't exactly
the same which was a bit confusing.