DEV: Refactor duplicate message blocker into class
We introduced duplicate message blocking in d1597107d2c0d2dbd18ce96ef71f82dd3d004063
but put the blocker logic which is quite specific inside the
ChatMessage model, and without many specs. This commit refactors
the methods and logic into a DuplicateMessageValidator class
and adds additional specs.

The numbers used for the calculations are quite mysterious, but
hopefully the specs cover the common use cases. The default for the
setting is 0.5, min is 0, and the max is 1.0