FIX: Onebox should only show active users
A onebox for a chat channel should only show users that are active, not
staged, and not suspended in the members list.

Maybe as a future feature we should boot users from the channel list that don't
belong to a certain group or that are no longer activated. Not really
sure, it's possible they should re-join the group if they are no longer
suspended.