FIX: Use resolveShortUrls for chat uploads
In core, when adding data-orig-src and data-orig-href to
uploads, we use resolveShortUrls when updating the composer
preview to retrieve the real upload URL and add it as the
src/href to the img/a tag. This commit does the same for chat
in a message decorator, which fixes the issue where the non-image
files did not have a URL to download.

Also removes the src from the image uploads in favour of using
resolveShortUrls for them too. The function is already debounced
and the found URLs cached, so this should not cause a performance
issue theoretically.