FIX: Chat quote interfering with oneboxes
A oneline change. Chat quotes render their inner content with the
chat markdown engine with a limited subset of rules, using a
html_raw markdown-it token. However when I first added this I
pushed the token to the state with the wrong level. I had this:

```
// rendering chat message content with limited markdown rule subset
const token = state.push("html_raw", "", 0);
token.content = customMarkdownCookFn(content);
state.push("html_raw", "", -1);
```

But needed this:

```
// rendering chat message content with limited markdown rule subset
const token = state.push("html_raw", "", 1);
token.content = customMarkdownCookFn(content);
state.push("html_raw", "", -1);
```

Spot the difference? (I simply changed the 0 to 1 on the first line).
This was causing the token "level" to go into negatives, which then
broke our onebox.js logic in core. If the previous token level is 0,
then we do a big onebox, otherwise we do an inline onebox because
we assume the link is lower in the HTML heirarchy:

https://github.com/discourse/discourse/blob/0ae7b43018238a883326da950f1e3bbf7e1f87a6/app/assets/javascripts/pretty-text/engines/discourse-markdown/onebox.js#L21-L22

So the level was -1 after a chat quote since the html_raw token
was pushed incorrectly, which meant that any oneboxes after the quote
would render inline.