SECURITY: Ensure chat message lookup checks for channel access
This commit primarily fixes a security issue where permissions were not
checked during message lookup.

In addition, a review of the ChatChannelsController and ChatController was
conducted to make sure there was not more holes. They have been
patched, along with the following changes made:

* Refactors ChatBaseController.set_channel_and_chatable_with_access_check to
  use DiscourseChat::ChatChannelFetcher.find_with_access_check directly, and
folds in the different code from the base controller into the fetcher for
consistency.
* Changes the POST /chat/drafts endpoint to use a chat_channel_id parameter
  instead of channel_id for consistency, and ensures channel access is correct
before creating the draft.
* Makes sure many endpoints missing the
  set_channel_and_chatable_with_access_check now have that check, both in the
ChatController and the ChatChannelsController.
* Also deletes some dead code and adds notes.