FEATURE: Automatically add users to a channel.
When creating a category channel, you have the option to configure it to add users automatically, which means all group members that have access to that category will join. Also, you can later enable or disable this option when
editing a channel.

Additionally, users will join the channel after gaining access to the category, or removed after losing it.