FIX: Allow user to direct message themself again
At some time in the past few months the ability for
users to create a new DM channel with themself was
broken, this commit fixes the issue by de-duplicating
the users that we are creating the channel for.

This commit also splits the create! method argument
for DirectMessageChannelCreator into acting_user/
target_user arguments, which will make it easier to
add mute/ignore checks involving the user creating
the message and who they are targeting.