FIX: Allow user to direct message themself again and fix DM inconsistencies (#990)
At some time in the past few months (possibly #758)
the ability for users to create a new DM channel with themselves
was broken. There was inconsistency around the
chat direct message controller sometimes accepting
an array of usernames and sometimes accepting
a comma separated string of usernames. They are
both needed - the former for POST and the latter
for GET. The test for POST in ruby did not pass the
usernames as an array so it was not catching the
error which was preventing the creation of the channel,
which was caused by a duplication of the current user
in the users array.

This commit also changes the `create!` method argument
for `DirectMessageChannelCreator` to have a
`target_user` argument, which will make it easier to
add mute/ignore checks involving the user creating
the message and who they are targeting, since an `acting_user`
argument will be needed for that change.

It also addresses some inconsistencies with how the usernames
are passed to the controller from JS, and adds missing specs
for `DirectMessageChannelCreator`