FIX: Use as_json instead of to_json for chat DM message bus
Follow up to f1fb1d0b4843b6265a6f46d8b24c31930a69c015,
in this commit I accidentally changed as_json (which produces
a Hash) to to_json (which produces a JSON string) in the
ChatPublisher for the new DM channel. This makes the MessageBus
data unparseable, so in JS in the chat service it ends up creating
a public channel in the channel list with undefined data, leaving
a blank channel in the list with a blue unread indicator.