FIX: Create channel modal error with type param required
Follow up to ca528966dfc4be9a920a8477813892e19a38d0a2, channel
creation was prevented because we still required the `type`
parameter in the controller but that param is redundant since
the only type of channel we can make now is a Category channel.
This commit fixes the issue and removes the type requirement.