FIX: Do not fetch messages for draft channel with no ID (#1000)
When starting a new DM channel with one or more users where
no channel exists we return a 404 from the server. However
we were still trying to fetch messages for that channel in
the chat-live-pane, leading to a GET like this:

http://localhost:4200/chat/undefined/messages.json?page_size=50

This commit fixes the issue by not doing any additional fetching
in chat-live-pane if the chat channel ID is not present.