FEATURE: Automatically add users to a channel. (#988)
* FEATURE: Automatically add users to a channel.

When creating a category channel, you have the option to configure it to add users automatically, which means all group members that have access to that category will join. Also, you can later enable or disable this option when
editing a channel.

Additionally, users will join the channel after gaining access to the category, or removed after losing it.

* Manage members when category permissions change

* Include join reason and seed default channels

* We don't support creating topic channels anymore

* Address code review feedback

* Rely on category permissions enforced by our guardian instead of trying to update all memberships' following attribute to false when they lose access. It simplifies the code and reduces the risk of flooding the queue with repeated
permission changes over a short time.

* Hide the auto join option is the max_auto_joined_users is set to 0

* Improve hints and display a confirmation dialog with the number of users who will automatically join the channel.

Make sure anonymized users and bots are excluded in the auto-join logic.

* Make sure auto-join works for both public and read-restricted categories.

Don't try to update the membership if it exists, but following is false. The user left the channel, so there's no need to add them again. Also, don't set it to false when the user gets removed from a group. Rely on category
permissions instead.

* Filter stageg and unactive users in the auto-join logic

* Publish updates to auto-joined users

* Make sure new users are auto-joined to channels.

Always use the async jobs for this process to ensure we apply the permission rules and send the correct message bus' messages.